refactor(organon): extract energeia::dispatch submodule (#2642)
Move dromeus tool definition and DromeusExecutor to dispatch.rs.
